Unary operators have problems when using the standard REFLECTION_REFLECTABLE_ADD_MEMBER_FUNCTION_REFLECTOR.
The problem seems to be in the variadic parameter pack. Although it will always be passed empty, the compiler fails at parsing time (it should simply SFINAE it away at instantiation time, *if* it was called with more parameters).

Created a special macro that only stamps out a non-templated version of the macro.