Fixed an issue in downcasting to the derived type.
Because all the reflectors inherit from the exact same base type, it is not possible to properly cast to the derived type, since the base type doesn't know via which reflector class it should cast in order to find the derived type (i.e. the using_<> class or derivative thereof).

- Made the derived member-function a free-function template, such that the reflector class is known inside the derived function.