Unifying customization points for member types and functions: on_call() / call() now also pass the explicit arguments via a normal parameter.
This is not strictly necessary for the member-function machinery, because there are fallback on_call() / call() functions defined. However, if we want to unify the mechanisms for member types and member functions, we need to get rid of these fallback functions, making the fix from the previous commit (d8fb339134aeabcdafa8f7fe8fc7eb70e09a7caf) necessary.