Removed invalid checks in unit tests.
This function will always select the const member-function overload, and therefore doesn't check whether the non-const won't be added.

Anyway, it might be up to the on_call hook to decide how this should be handled. So let's leave it for now, until we discover a real problem that needs fixing.