Added fix for preventing operator=() from being added to the overload set while compiling the copy assignment operators.
Probably the problem is as follows:
- While defining the class hierarchy, starting from the smart reference class (e.g. Property<int>), the exposer class for the assignment operator is being defined.
- While defining that exposer class, the copy-assignment operator gets instantiated for that class.
- While considering all the assignment operators for that class, the custom 'smart reference' operator= gets instantiated as well, which results in the following cascade of instantiations:
- on_call gets instantiated
  - which tries to extract the delegate object out of the smart reference (the self parameter)
  - this fails, because the smart reference class is at that point still an incomplete type, such that the delegate_if_is_using() function falls back to an identity function
- Now, on_call, calls the 'call' function using the 'self' parameter as-is (i.e. the delegate object has *not* been extracted).
- The 'call' function in turn, performs the assignment operation, which in turn calls the custom 'smart reference' operator=, because the left-hand-side object is the smart reference class, i.e. the same type that we started with.

Although I don't understand why suddenly it is possible why the operator= can act on the smart reference class (e.g. Property<int>), considering it was previously an incomplete type, it seems that this problem results in an infinite recursive loop *at compile time*.

The fix for this is to make the custom 'smart reference' operator= fail if it is being instantiated for any argument (i.e. the right-hand-side's operand) that is one of the classes in the inheritance hierarchy between the smart reference class and the exposer class. Note that it is not possible to use an is_base_of check, because the relationship between the classes cannot be queried for incomplete types. Instead, we keep track of a list of types which represent the inheritance hierarchy.