Removed unnecessary and incorrect static_cast from free function reflector, which would trigger undefined behavior.
In the following code:
auto x = Property<int>{5};

auto y = x + 1;

cout << y << endl; // prints garbage, e.g. -448374440

the '1' would be converted into a const Property<int> &, because of the static_cast_if_possible. In this case, it was possible, but results in a reference to a local (inside of the static_cast_if_possible_impl function). This in turn triggers undefined behavior, observed by the garbage output.