Added super ugly hack that might fix the adl problem for uniform call syntax.
the problem is as follows:

using namespace algorithms;
$(v).sorted();
     \-> on_call() -> call() -> sorted(v)
}
the last sorted() function doesn't know which namespaces it should consider, because what has been inserted in the caller's scope (i.e. using namespace algorithms;) is lost there.

The idea is to register all the namespaces that need to be considered, by inserting a tag type into each of them. Then an additional smart reference is wrapped around the underlying object before calling the actual free-function (i.e. sorted(v)). This smart reference contains as a template argument the inserted tag type. As a result, adl will also consider the tag type's namespace, which is the namespace that contains the correct function. Because the smart reference type implicitly converts to the underlying type, everything works automagically.