Turned adl_finder into a full-fledged smart reference.
This is needed in case the member function or free function that is being called captures the parameter as a templatized type, in which case the underlying object will not be pulled out of the adl_finder. In that case, it is very likely that the called function uses the interface of the passed parameter (which is now a smart reference). Therefore, it needs to forward the entire interface, and not only the conversion to the underlying type. Hence, a full-fledged smart reference type is required.